# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoTM API",
    "version": "0.1.0",
    "description": "AutoTM Marketplace API — Phase 1"
  },
  "servers": [
    {
      "url": "https://api.auto.tm"
    }
  ],
...
  "paths": {
...
    "/api/v1/listings/{id}/inspection-interest": {
      "post": {
        "summary": "Register inspection interest (fake-door)",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "willingnessToPayTmt": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing interest returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "listingId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "requesterUserId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "side": {
                      "type": "string",
                      "enum": [
                        "buyer",
                        "seller"
                      ]
                    },
                    "willingnessToPayTmt": {
                      "type": "integer",
                      "nullable": true,
                      "minimum": 0,
                      "maximum": 10000
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "reusedExisting": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "listingId",
                    "requesterUserId",
                    "side",
                    "willingnessToPayTmt",
                    "createdAt",
                    "updatedAt",
                    "reusedExisting"
                  ]
                }
              }
            }
          },
          "201": {
            "description": "Interest created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "listingId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "requesterUserId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "side": {
                      "type": "string",
                      "enum": [
                        "buyer",
                        "seller"
                      ]
                    },
                    "willingnessToPayTmt": {
                      "type": "integer",
                      "nullable": true,
                      "minimum": 0,
                      "maximum": 10000
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "reusedExisting": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "listingId",
                    "requesterUserId",
                    "side",
                    "willingnessToPayTmt",
                    "createdAt",
                    "updatedAt",
                    "reusedExisting"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_FAILED",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "nullable": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message",
                    "timestamp",
                    "requestId"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Feature disabled or user suspended",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_FAILED",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "nullable": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message",
                    "timestamp",
                    "requestId"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Listing not found or ineligible",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_FAILED",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "nullable": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message",
                    "timestamp",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/inspection-interests": {
      "get": {
        "summary": "List inspection interest statistics",
        "tags": [
          "Admin",
          "Reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 10,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Interest statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "listingId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "totalInterest": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "buyerInterest": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "sellerInterest": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "willingnessToPayTmtSum": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "willingnessToPayTmtCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "willingnessToPayTmtAvg": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "listingId",
                          "totalInterest",
                          "buyerInterest",
                          "sellerInterest",
                          "willingnessToPayTmtSum",
                          "willingnessToPayTmtCount",
                          "willingnessToPayTmtAvg"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "page": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "pageSize": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "items",
                    "total",
                    "page",
                    "pageSize",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_FAILED",
                        "UNAUTHORIZED",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "RATE_LIMITED",
                        "INTERNAL"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "nullable": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message",
                    "timestamp",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}